+++ Files.Launcher/FilePermissions.cs
@@ -1,7 +1,9 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.IO;
 using System.Linq;
 using System.Security.AccessControl;
+using System.Security.Principal;
 using System.Text;
 using System.Threading.Tasks;
 
@@ -19,6 +21,33 @@ namespace FilesFullTrust
         }
 
         public List<FileSystemAccessRule2> AccessRules { get; set; }
+
+        public static FilePermissions FromFilePath(string filePath, bool isFolder)
+        {
+            var filePermissions = new FilePermissions() { Path = filePath };
+            try
+            {
+                if (isFolder && Directory.Exists(filePath))
+                {
+                    var acs = Directory.GetAccessControl(filePath);
+                    var accessRules = acs.GetAccessRules(true, true, typeof(SecurityIdentifier));
+                    filePermissions.AccessRules.AddRange(accessRules.Cast<FileSystemAccessRule>().Select(x => FileSystemAccessRule2.FromFileSystemAccessRule(x)));
+                }
+                else if (File.Exists(filePath))
+                {
+                    var acs = File.GetAccessControl(filePath);
+                    var accessRules = acs.GetAccessRules(true, true, typeof(SecurityIdentifier));
+                    filePermissions.AccessRules.AddRange(accessRules.Cast<FileSystemAccessRule>().Select(x => FileSystemAccessRule2.FromFileSystemAccessRule(x)));
+                }
+                filePermissions.CanReadFilePermissions = true;
+            }
+            catch (UnauthorizedAccessException)
+            {
+                // User does not have rights to read access rules
+                filePermissions.CanReadFilePermissions = false;
+            }
+            return filePermissions;
+        }
     }
 
     public class FileSystemAccessRule2
